Import some of the TeX notes into Markdown+KaTeX.
Notes on how this hell process works:

This is kind of terrible because there's a bunch of stuff that KaTeX doesn't
support (e.g., macros with arguments), so it's a real pain to import the notes.

To do this, start with pandoc:

    pandoc -f latex -t markdown input.tex -o output.md

Pandoc uses $...$ for inline math and $$...$$ for display math, but we want
\(...\) and \[...\]. This means we actually need \\( and \\) in the Markdown in
order to get \( \) in the generated HTML.  To do this conversion, try (in vim)

     :%s/\$\$\(\_.\{-}\)\$\$/\\\\[\r\1\r\\\\]/gc

to change $$...$$ to \[ newline ... newline \]. Note that \_ in the match means
"match across lines" and \{-} means "*, but non-greedy", while \r is used in
the substitution for a newline because Vim thinks \n should mean null byte (but
only when substituting, not when matching).

Next replace $s with

     :%s/\$\(\_.\{-}[^\\]\)\$/\\\\(\1\\\\)/gc

the extra [^\\] is to prevent matching \$ as an endpoint of a $...$.

Finally, fix up errors.