Add and test Serde support.
This changes the rangeproof test code to explicitly pass a proof from the
prover to the verifier.

The test code uses bincode for serialization, but the generated proof sizes are
quite bloated (up to 200 bytes larger than necessary).

Some points:

1. Ser/deser is connected with proof verification, since we want to feed the
prover's commitments into the hash.  Right now we decompress and recompress, so
it would probably be better to have the Proof struct hold compressed points.

However, the serde support in `curve25519-dalek` is designed so that the input
validation (are points valid? are scalars canonically encoded?) happens
automatically during deserialization.  The compressed point formats don't
implement `Serialize` or `Deserialize`, so that structs with points and scalars
are either all valid or just return an error.

2. We could write custom Serialize/Deserialize implementations, but it would be
good to have custom types for each resolution of range proof so that we don't
need to encode lengths.

We also would have to make sure that repeat all of the checks correctly.

3. Bincode is designed for IPC, so its goal is just encoding <= memory size,
and I think it encodes lengths as u64s, for instance.  So something like CBOR
might do better.

I feel like the best thing to do is to specialize the `RangeProof` struct by
bitsize, and then provide custom `Serialize`/`Deserialize` implementations that
work on a fixed-size byte array.  This would give compact proof sizes with any
Serde backend.