Move transcript state inside of the Dealer state
Previously, the transcript state was maintained by the protocol orchestrator,
and passed in as a parameter to each step of the dealer's work. This changes
the code so that the dealer holds on to the transcript reference until the
protocol is complete.

This means

a) it's one less parameter
b) it would disallow any other code from touching the transcript state while
the dealer is active.

b) is kind of a fringe benefit, but previously, it's conceivably possible to
write code that fed an extra message into the transcript between dealer steps,
which would give an invalid proof (since the challenges would not replay).