Pass generators into the aggregated proof code and shorten proofs
This saves a ton of work building the generators, and makes the proofs smaller
by dropping the value commitments and the bitsize/number of parties.

Actually using the `SinglePartyAggregator` is a bit awkward, since it requires
creating all of the value commitments by hand. I'm not sure what the best
solution to that is yet.

Progress towards #74