Make the ConstraintSystem keep knowledge of the blinding factors.
The ConstraintSystem constructor for the prover takes a vector of blinding
factors, which it uses to build the `V` points (commitments to the input wires)
that it adds into the transcript.

Later, the proving code requires knowledge of the blinding factors.  This
change makes the constraint system hold on to the blinding factors for the
input wires.  This simplifies the proving API (since the prover can pass
ownership of the blinding factor vector into the `ConstraintSystem`) and will
also allow the constraint system to construct a `TranscriptRng` with entropy
from the blinding factors.