Add a conversion from ProofError to R1CSError.
In the long term, I think that it would be better to combine R1CSError with
ProofError, if we can do that without breaking stability.  If not, we should
refactor the TranscriptProtocol so that it returns a TranscriptError type that
can be converted into either a ProofError or an R1CSError.